requirements.txt
astroid==3.3.8
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
Flask==3.1.0
idna==3.10
injector==0.22.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pylint==3.3.4
python-dotenv==1.0.1
requests==2.32.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
